Fix bounds
